+++ src/errors/HttpError.ts
@@ -1,9 +1,11 @@
 export class HttpError extends Error {
   readonly statusCode: number;
+  readonly code: string;
 
-  constructor(statusCode: number, message: string) {
+  constructor(statusCode: number, message: string, code = "http_error") {
     super(message);
     this.name = "HttpError";
     this.statusCode = statusCode;
+    this.code = code;
   }
 }